added content loading animation
