+++ src/styles/constants.css
@@ -5,4 +5,5 @@
   --pink: #ffa7c4;
   --pink-hover: #f082ac;
   --comment-meta-color: hsla(0, 0%, 100%, 0.88);
+  --light-animated-text: hsla(0, 0%, 100%, 0.88);
 }
